blog-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic-settings==2.1.0
python-slugify==8.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.0
